requirements.txt
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
google==1.9.3
gps==3.16
html5lib==0.999999999
idna==2.5
image==1.5.7
olefile==0.44
Pillow==4.2.1
PyExifTool==0.1
pygobject==3.24.1
pytz==2017.2
requests==2.18.1
six==1.10.0
urllib3==1.21.1
vboxapi==1.0
webencodings==0.5.1
